+++ metadata.go
@@ -7,7 +7,7 @@ type Metadata struct {
 	Subtype    string                     `json:"subtype,omitempty"`
 	Type       string                     `json:"type"`
 	Value      any                        `json:"value"`
-	Visibility []MetadataVisibilityOption `json:"visibility"`
+	Visibility []MetadataVisibilityOption `json:"visibility,omitempty"`
 }
 
 type MetadataVisibilityOption string
@@ -22,3 +22,13 @@ type MetadataMatchedResult struct {
 	ID       string     `json:"id"`
 	Metadata []Metadata `json:"metadata"`
 }
+
+type MetadataQueryKey = string
+
+const (
+	MetadataQueryKeyAdmins       MetadataQueryKey = "admins"
+	MetadataQueryKeyDestinations MetadataQueryKey = "destinations"
+	MetadataQueryKeyRecipients   MetadataQueryKey = "recipients"
+	MetadataQueryKeyTasks        MetadataQueryKey = "tasks"
+	MetadataQueryKeyWorkers      MetadataQueryKey = "workers"
+)