+++ service/task/client.go
@@ -4,17 +4,17 @@ import (
 	"net/http"
 
 	onfleet "github.com/onfleet/gonfleet"
-	"github.com/onfleet/gonfleet/netw"
+	"github.com/onfleet/gonfleet/netwrk"
 )
 
 type Client struct {
 	apiKey       string
-	rlHttpClient *netw.RlHttpClient
+	rlHttpClient *netwrk.RlHttpClient
 	url          string
-	call         netw.Caller
+	call         netwrk.Caller
 }
 
-func Plug(apiKey string, rlHttpClient *netw.RlHttpClient, url string, call netw.Caller) *Client {
+func Plug(apiKey string, rlHttpClient *netwrk.RlHttpClient, url string, call netwrk.Caller) *Client {
 	return &Client{
 		apiKey:       apiKey,
 		rlHttpClient: rlHttpClient,
@@ -23,6 +23,7 @@ func Plug(apiKey string, rlHttpClient *netw.RlHttpClient, url string, call netw.
 	}
 }
 
+// Reference https://docs.onfleet.com/reference/get-single-task
 func (c *Client) Get(taskId string) (onfleet.Task, error) {
 	task := onfleet.Task{}
 	err := c.call(
@@ -38,6 +39,7 @@ func (c *Client) Get(taskId string) (onfleet.Task, error) {
 	return task, err
 }
 
+// Reference https://docs.onfleet.com/reference/get-single-task-by-shortid
 func (c *Client) GetByShortId(taskShortId string) (onfleet.Task, error) {
 	task := onfleet.Task{}
 	err := c.call(
@@ -55,3 +57,21 @@ func (c *Client) GetByShortId(taskShortId string) (onfleet.Task, error) {
 
 func (c *Client) ListTasks(teamId string) {
 }
+
+// Reference https://docs.onfleet.com/reference/querying-by-metadata
+func (c *Client) ListWithMetadataQuery(metadata []onfleet.Metadata) ([]onfleet.Task, error) {
+	tasks := []onfleet.Task{}
+	err := c.call(
+		c.apiKey,
+		c.rlHttpClient,
+		http.MethodPost,
+		c.url,
+		[]string{"metadata"},
+		nil,
+		metadata,
+		&tasks,
+	)
+	return tasks, err
+}
+
+// Reference