+++ service/team/client.go
@@ -4,17 +4,17 @@ import (
 	"net/http"
 
 	"github.com/onfleet/gonfleet"
-	"github.com/onfleet/gonfleet/netw"
+	"github.com/onfleet/gonfleet/netwrk"
 )
 
 type Client struct {
 	apiKey       string
-	rlHttpClient *netw.RlHttpClient
+	rlHttpClient *netwrk.RlHttpClient
 	url          string
-	call         netw.Caller
+	call         netwrk.Caller
 }
 
-func Plug(apiKey string, rlHttpClient *netw.RlHttpClient, url string, call netw.Caller) *Client {
+func Plug(apiKey string, rlHttpClient *netwrk.RlHttpClient, url string, call netwrk.Caller) *Client {
 	return &Client{
 		apiKey:       apiKey,
 		rlHttpClient: rlHttpClient,